+++ src/mojang/api/hash.hpp
@@ -33,7 +33,7 @@ namespace mojang::api {
 
         static void to_hex(hashed_array data, char* hex) {
             constexpr char hex_map[] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
-            for (int i = 0; i < DIGEST_LENGTH; i++) {
+            for (size_t i = 0; i < DIGEST_LENGTH; i++) {
                 hex[i * 2] = hex_map[(data[i] & 0xf0) >> 4];
                 hex[i * 2 + 1] = hex_map[data[i] & 0x0f];
             }
